+++ projects/recalbox-hardware/case/installers/nuxii/install.py
@@ -1,0 +1,27 @@
+from installers.base.install import InstallBase
+
+
+class Install(InstallBase):
+
+    def __init__(self):
+        InstallBase.__init__(self)
+
+    def InstallHardware(self, case):
+
+        return False
+
+    def InstallSoftware(self, case):
+
+        return case
+
+    def UninstallHardware(self, case):
+
+        return False
+
+    def UninstallSoftware(self, case):
+
+        return ""
+
+    def GetInstallScript(self, case):
+
+        return None